pyproject.toml
[tool.poetry]
name = "bld"
version = "0.1.0"
description = ""
authors = ["yusuf-74 <yachour31@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.2"
djangorestframework = "^3.14.0"
Markdown = "^3.4.1"
django-filter = "^22.1"


[tool.poetry.group.dev.dependencies]
black = {version = "^22.10.0", allow-prereleases = true}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
